+++ src/org/csc133/a1/Helicopter.java
@@ -5,7 +5,7 @@ public class Helicopter extends MovableObject implements ISteerable
     private int stickAngle;
     private int maximumSpeed;
     private int fuelLevel;
-    private float fuelConsumptionRate;
+    private double fuelConsumptionRate;
     private int damageLevel;
     private int lastSkyscraperReached;
 
@@ -15,7 +15,7 @@ public class Helicopter extends MovableObject implements ISteerable
         this(25, 0, 0, 150, 45, 0, 45, 10, 100, 5, 0, 0);
     }
 
-    public Helicopter(int size, float x, float y, int color, int heading, int speed, int stickAngle, int maximumSpeed, int fuelLevel, float fuelConsumptionRate, int damageLevel, int lastSkyscraperReached)
+    public Helicopter(int size, double x, double y, int color, int heading, int speed, int stickAngle, int maximumSpeed, int fuelLevel, double fuelConsumptionRate, int damageLevel, int lastSkyscraperReached)
     {
         super(size, x, y, color, heading, speed);
         this.stickAngle = stickAngle;
